Product Authentication service of customer mobile RFID devices
A product  authentication  service  which    is    anti-counterfeiting
technology and can be  applied to both the  supply chain (SCM) and
forthcoming  mobile RFID environment.  It aims to become  aware  of
distribution of  spurious  products  with  fake  RFID  tags and  to
provide a  product authentication  service  to  consumers with mobile
RFID  devices  like   mobile  phone  which has a  RFID reader. We  also
propose  an authentication  mechanism between  a  tag  and an
information  server through the device.  The  mechanism is designed to
be  appropriated  for  EPC Class-1  Generation-2  UHF tags. The tags,
one of  RFID  tag  types,  is expected  that  most  companies are
planning  to use  it  in  the  SCM  in  the short term  and  in consumer
packaging  in  the  long  term  due  to  its  inexpensive cost. Because
of the very  cost,   however,  its  resources   are  extremely scarce
and  it is  hard to  have any  valuable  security algorithms   in  it.
It  causes   security  vulnerabilities, in  particular  cloning  the
tags  for counterfeits.  The  mechanism  provides  secure  mutual
authentication  between  a   tag  and a server. It can be  applied to
the mobile RFID  environment that any mobile RFID  reader  can  be
placed  in the interface  between the tag and the server. It means that
the mechanism provides secure  communication  between them  even  the
reader  is  not trusted. On  basis  of  this  tag authentication, the
service makes use of  location  information  especially  from  LBS
(location based service) of  the consumers ' devices.